+++ projects/_03_independent_cqc/_04_model/tests/utils/test_validate_models.py
@@ -14,8 +14,8 @@ from projects._03_independent_cqc.unittest_data.polars_ind_cqc_test_file_schemas
 class TestGetExpectedRowCount(unittest.TestCase):
     def setUp(self):
         self.df_test = pl.DataFrame(
-            Data.non_res_with_dormancy_rows,
-            Schemas.non_res_with_dormancy_schema,
+            Data.validate_model_feature_rows,
+            Schemas.validate_model_feature_schema,
             orient="row",
         )
 
@@ -33,5 +33,14 @@ class TestGetExpectedRowCount(unittest.TestCase):
         expected_row_count = Data.expected_get_expected_row_count_rows
         returned_row_count = job.get_expected_row_count_for_model_features(
             self.df_test, "non_res_with_dormancy_model"
+        )
+        self.assertEqual(expected_row_count, returned_row_count)
+
+    def test_get_expected_row_count_for_care_home_model(
+        self,
+    ):
+        expected_row_count = Data.expected_get_expected_row_count_rows
+        returned_row_count = job.get_expected_row_count_for_model_features(
+            self.df_test, "care_home_model"
         )
         self.assertEqual(expected_row_count, returned_row_count)